- Some refactoring  - The cover image in the new language  - A few bug fixes  - Materials and transforms can be extended
