+++ Parser/LanguageParser.Surfaces.cs
@@ -16,627 +16,6 @@ namespace RayTracer.Parser;
 public partial class LanguageParser
 {
     /// <summary>
-    /// This method is used to handle the beginning of a plane block.
-    /// </summary>
-    private void HandleStartPlaneClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Plane");
-
-        PlaneInstructionSet instructionSet = ParsePlaneClause();
-
-        _ = new TopLevelObjectInstruction<Plane>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a plane block.
-    /// </summary>
-    private PlaneInstructionSet ParsePlaneClause()
-    {
-        PlaneInstructionSet instructionSet = new ();
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandlePlaneEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a plane block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandlePlaneEntryClause(Clause clause)
-    {
-        PlaneInstructionSet instructionSet = (PlaneInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "plane");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a sphere block.
-    /// </summary>
-    private void HandleStartSphereClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Sphere");
-
-        SphereInstructionSet instructionSet = ParseSphereClause();
-
-        _ = new TopLevelObjectInstruction<Sphere>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a sphere block.
-    /// </summary>
-    private SphereInstructionSet ParseSphereClause()
-    {
-        SphereInstructionSet instructionSet = new ();
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandleSphereEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a sphere block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleSphereEntryClause(Clause clause)
-    {
-        SphereInstructionSet instructionSet = (SphereInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "sphere");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a cube block.
-    /// </summary>
-    private void HandleStartCubeClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Cube");
-
-        CubeInstructionSet instructionSet = ParseCubeClause();
-
-        _ = new TopLevelObjectInstruction<Cube>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a cube block.
-    /// </summary>
-    private CubeInstructionSet ParseCubeClause()
-    {
-        CubeInstructionSet instructionSet = new ();
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandleCubeEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a cube block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleCubeEntryClause(Clause clause)
-    {
-        CubeInstructionSet instructionSet = (CubeInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "cube");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a circular surface block.
-    /// </summary>
-    private void HandleStartCircularSurfaceClause(Clause clause)
-    {
-        bool open = clause.Tokens[0].Text == "open";
-
-        if (open)
-            clause.Tokens.RemoveFirst();
-
-        string text = clause.Tokens[0].Text;
-
-        VerifyDefaultSceneUsage(clause, $"{char.ToUpper(text[0])}{text[1..]}");
-
-        switch (text)
-        {
-            case "cylinder":
-            {
-                CylinderInstructionSet cylinderInstructionSet = ParseCylinderClause(open);
-                _ = new TopLevelObjectInstruction<Cylinder>(_context.InstructionContext, cylinderInstructionSet);
-                break;
-            }
-            case "conic":
-            {
-                ConicInstructionSet conicInstructionSet = ParseConicClause(open);
-                _ = new TopLevelObjectInstruction<Conic>(_context.InstructionContext, conicInstructionSet);
-                break;
-            }
-            default:
-                throw new TokenException($"Internal error: unknown circular surface type: {text}.")
-                {
-                    Token = clause.Tokens[0]
-                };
-        }
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a cylinder block.
-    /// </summary>
-    private CylinderInstructionSet ParseCylinderClause(bool open)
-    {
-        CylinderInstructionSet instructionSet = new ();
-
-        ParseCircularSurfaceClause(instructionSet, open);
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a conic block.
-    /// </summary>
-    private ConicInstructionSet ParseConicClause(bool open)
-    {
-        ConicInstructionSet instructionSet = new ();
-
-        ParseCircularSurfaceClause(instructionSet, open);
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a circularSurface block.
-    /// </summary>
-    private void ParseCircularSurfaceClause<TObject>(
-        ObjectInstructionSet<TObject> instructionSet, bool open)
-        where TObject : CircularSurface, new()
-    {
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("circularSurfaceEntryClause", HandleCircularSurfaceEntryClause);
-
-        _context.PopInstructionSet();
-
-        if (open)
-        {
-            instructionSet.AddInstruction(new SetObjectPropertyInstruction<TObject, bool>(
-                target => target.Closed, false));
-        }
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a circularSurface block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleCircularSurfaceEntryClause(Clause clause)
-    {
-        IInstructionSet instructionSet = _context.CurrentSet;
-
-        switch (instructionSet)
-        {
-            case CylinderInstructionSet cylinderInstructionSet:
-                HandleCircularSurfaceEntryClause(cylinderInstructionSet, clause);
-                break;
-            case ConicInstructionSet conicInstructionSet:
-                HandleCircularSurfaceEntryClause(conicInstructionSet, clause);
-                break;
-            default:
-                throw new Exception("Internal error: unknown circular surface instruction set type.");
-        }
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a circularSurface block.
-    /// </summary>
-    /// <param name="instructionSet">The instruction set to work with.</param>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleCircularSurfaceEntryClause<TObject>(
-        ObjectInstructionSet<TObject> instructionSet, Clause clause)
-        where TObject : CircularSurface, new()
-    {
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-        {
-            string text = clause.Tokens[0].Text;
-
-            switch (text)
-            {
-                case "min":
-                    instructionSet.AddInstruction(new SetObjectPropertyInstruction<TObject, double>(
-                        target => target.MinimumY, (Term) clause.Expressions[0]));
-                    break;
-                case "max":
-                    instructionSet.AddInstruction(new SetObjectPropertyInstruction<TObject, double>(
-                        target => target.MaximumY, (Term) clause.Expressions[0]));
-                    break;
-                default:
-                    HandleSurfaceClause(clause, instructionSet, "circularSurface");
-                    break;
-            }
-        }
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a triangle block.
-    /// </summary>
-    private void HandleStartTriangleClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Triangle");
-
-        TriangleInstructionSet instructionSet = ParseTriangleClause(clause);
-
-        _ = new TopLevelObjectInstruction<Triangle>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a triangle block.
-    /// </summary>
-    private TriangleInstructionSet ParseTriangleClause(Clause clause)
-    {
-        Term first = (Term) clause.Expressions[0];
-        Term second = (Term) clause.Expressions[1];
-        Term third = (Term) clause.Expressions[2];
-
-        TriangleInstructionSet instructionSet = new (first, second, third);
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandleTriangleEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a triangle block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleTriangleEntryClause(Clause clause)
-    {
-        TriangleInstructionSet instructionSet = (TriangleInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "triangle");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a smooth triangle block.
-    /// </summary>
-    private void HandleStartSmoothTriangleClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Smooth triangle");
-
-        SmoothTriangleInstructionSet instructionSet = ParseSmoothTriangleClause(clause);
-
-        _ = new TopLevelObjectInstruction<SmoothTriangle>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a smoothTriangle block.
-    /// </summary>
-    private SmoothTriangleInstructionSet ParseSmoothTriangleClause(Clause clause)
-    {
-        List<Term> terms = clause.Expressions
-            .Cast<Term>()
-            .ToList();
-
-        SmoothTriangleInstructionSet instructionSet = new (
-            terms[0], terms[1], terms[2],
-            terms[3], terms[4], terms[5]);
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandleSmoothTriangleEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a smoothTriangle block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleSmoothTriangleEntryClause(Clause clause)
-    {
-        SmoothTriangleInstructionSet instructionSet = (SmoothTriangleInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "smoothTriangle");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a object file block.
-    /// </summary>
-    private void HandleStartObjectFileClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Object file");
-
-        ObjectFileInstructionSet instructionSet = ParseObjectFileClause(clause);
-
-        _ = new TopLevelObjectInstruction<Group>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a object file block.
-    /// </summary>
-    private ObjectFileInstructionSet ParseObjectFileClause(Clause clause)
-    {
-        Term fileName = (Term) clause.Expressions[0];
-
-        ObjectFileInstructionSet instructionSet = new (CurrentDirectory, fileName);
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("surfaceEntryClause", HandleObjectFileEntryClause);
-
-        _context.PopInstructionSet();
-
-        instructionSet.AddInstruction(new FinalizeGroupInstruction());
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a objectFile block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleObjectFileEntryClause(Clause clause)
-    {
-        ObjectFileInstructionSet instructionSet = (ObjectFileInstructionSet) _context.CurrentSet;
-
-        if (clause == null) // We must have hit a transform property...
-            HandleSurfaceTransform(instructionSet);
-        else
-            HandleSurfaceClause(clause, instructionSet, "object file");
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a CSG block.
-    /// </summary>
-    private void HandleStartCsgClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, clause.Tokens[0].Text);
-
-        CsgSurfaceInstructionSet instructionSet = ParseCsgClause(clause);
-
-        _ = new TopLevelObjectInstruction<CsgSurface>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a csg block.
-    /// </summary>
-    /// <param name="clause">The clause that started the csg.</param>
-    private CsgSurfaceInstructionSet ParseCsgClause(Clause clause)
-    {
-        string text = clause.Tokens[0].Text;
-        CsgOperation operation = text switch
-        {
-            "union" => CsgOperation.Union,
-            "difference" => CsgOperation.Difference,
-            "intersection" => CsgOperation.Intersection,
-            _ => throw new Exception($"Internal error: unknown CSG type: {text}.")
-        };
-
-        CsgSurfaceInstructionSet instructionSet = new (operation, clause.Tokens[0]);
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("csgEntryClause", HandleCsgEntryClause);
-
-        _context.PopInstructionSet();
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a csg block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleCsgEntryClause(Clause clause)
-    {
-        CsgSurfaceInstructionSet instructionSet = (CsgSurfaceInstructionSet) _context.CurrentSet;
-
-        if (clause == null)
-        {
-            HandleSurfaceTransform(instructionSet);
-
-            return;            
-        }
-
-        switch (clause.Tag)
-        {
-            case "plane":
-                instructionSet.AddInstruction(ParsePlaneClause());
-                break;
-            case "sphere":
-                instructionSet.AddInstruction(ParseSphereClause());
-                break;
-            case "cube":
-                instructionSet.AddInstruction(ParseCubeClause());
-                break;
-            case "circularSurface":
-                bool open = clause.Tokens[0].Text == "open";
-                if (clause.Tokens[0].Text == "cylinder" || clause.Tokens[1].Text == "cylinder")
-                    instructionSet.AddInstruction(ParseCylinderClause(open));
-                else if (clause.Tokens[0].Text == "conic" || clause.Tokens[1].Text == "conic")
-                    instructionSet.AddInstruction(ParseConicClause(open));
-                else
-                    throw new Exception("Internal error: unknown circular surface type.");
-                break;
-            case "triangle":
-                instructionSet.AddInstruction(ParseTriangleClause(clause));
-                break;
-            case "smooth":
-                instructionSet.AddInstruction(ParseSmoothTriangleClause(clause));
-                break;
-            case "objectFile":
-                instructionSet.AddInstruction(ParseObjectFileClause(clause));
-                break;
-            case "csg":
-                instructionSet.AddInstruction(ParseCsgClause(clause));
-                break;
-            case "group":
-                instructionSet.AddInstruction(ParseGroupClause(clause));
-                break;
-            case "surface":
-                HandleSurfaceClause(clause, instructionSet, "csg");
-                break;
-        }
-    }
-
-    /// <summary>
-    /// This method is used to handle the beginning of a group block.
-    /// </summary>
-    private void HandleStartGroupClause(Clause clause)
-    {
-        VerifyDefaultSceneUsage(clause, "Group");
-
-        GroupInstructionSet instructionSet = ParseGroupClause(clause);
-
-        _ = new TopLevelObjectInstruction<Group>(_context.InstructionContext, instructionSet);
-    }
-
-    /// <summary>
-    /// This method is used to create the instruction set from a group block.
-    /// </summary>
-    /// <param name="clause">The clause that started the group.</param>
-    private GroupInstructionSet ParseGroupClause(Clause clause)
-    {
-        Token token = clause.Tokens[1];
-        string variableName = null;
-        Term startTerm = null;
-        Term endTerm = null;
-        Term stepTerm = null;
-        bool startIsOpen = false;
-        bool endIsOpen = false;
-
-        if (token is IdToken or KeywordToken)
-        {
-            clause.Tokens.RemoveRange(1, 2);
-
-            variableName = token.Text;
-            token = clause.Tokens[1];
-        }
-
-        if (BounderToken.LeftParen.Matches(token) ||
-            BounderToken.OpenBracket.Matches(token))
-        {
-            startTerm = (Term) clause.Expressions[0];
-            endTerm = (Term) clause.Expressions[1];
-            startIsOpen = BounderToken.LeftParen.Matches(token);
-            endIsOpen = BounderToken.LeftParen.Matches(clause.Tokens[3]);
-
-            clause.Expressions.RemoveRange(0, 2);
-            clause.Tokens.RemoveRange(1, 3);
-
-            if (clause.Tokens[1].Text == "by")
-            {
-                stepTerm = (Term) clause.Expressions[0];
-
-                clause.Expressions.RemoveFirst();
-                clause.Tokens.RemoveAt(1);
-            }
-        }
-
-        GroupInstructionSet instructionSet = new (
-            variableName, startTerm, endTerm, stepTerm, startIsOpen, endIsOpen);
-
-        _context.PushInstructionSet(instructionSet);
-
-        ParseBlock("groupEntryClause", HandleGroupEntryClause);
-
-        _context.PopInstructionSet();
-
-        instructionSet.AddInstruction(new FinalizeGroupInstruction());
-
-        return instructionSet;
-    }
-
-    /// <summary>
-    /// This method is used to handle an item clause of a group block.
-    /// </summary>
-    /// <param name="clause">The clause to process.</param>
-    private void HandleGroupEntryClause(Clause clause)
-    {
-        GroupInstructionSet instructionSet = (GroupInstructionSet) _context.CurrentSet;
-
-        if (clause == null)
-        {
-            HandleSurfaceTransform(instructionSet);
-
-            return;            
-        }
-
-        switch (clause.Tag)
-        {
-            case "plane":
-                instructionSet.AddInstruction(ParsePlaneClause());
-                break;
-            case "sphere":
-                instructionSet.AddInstruction(ParseSphereClause());
-                break;
-            case "cube":
-                instructionSet.AddInstruction(ParseCubeClause());
-                break;
-            case "circularSurface":
-                bool open = clause.Tokens[0].Text == "open";
-                if (clause.Tokens[0].Text == "cylinder" || clause.Tokens[1].Text == "cylinder")
-                    instructionSet.AddInstruction(ParseCylinderClause(open));
-                else if (clause.Tokens[0].Text == "conic" || clause.Tokens[1].Text == "conic")
-                    instructionSet.AddInstruction(ParseConicClause(open));
-                else
-                    throw new Exception("Internal error: unknown circular surface type.");
-                break;
-            case "triangle":
-                instructionSet.AddInstruction(ParseTriangleClause(clause));
-                break;
-            case "smooth":
-                instructionSet.AddInstruction(ParseSmoothTriangleClause(clause));
-                break;
-            case "objectFile":
-                instructionSet.AddInstruction(ParseObjectFileClause(clause));
-                break;
-            case "csg":
-                instructionSet.AddInstruction(ParseCsgClause(clause));
-                break;
-            case "group":
-                instructionSet.AddInstruction(ParseGroupClause(clause));
-                break;
-            case "boundingBox":
-                instructionSet.AddInstruction(new SetBoundingBoxInstruction(
-                    (Term) clause.Expressions[0], (Term) clause.Expressions[1]));
-                break;
-            case "surface":
-                HandleSurfaceClause(clause, instructionSet, "group");
-                break;
-        }
-    }
-
-    /// <summary>
     /// This method is used to handle a clause for general surface properties.
     /// </summary>
     /// <param name="clause">The clause to process.</param>
@@ -652,10 +31,8 @@ public partial class LanguageParser
         ObjectInstruction<TObject> instruction = field switch
         {
             "named" => CreateNamedInstruction<TObject>(term),
-            "material" => GetMaterialInstruction<TObject>(clause.Tokens[1]),
-            "transform" => new SetObjectPropertyInstruction<TObject, Matrix>(
-                target => target.Transform, new VariableTerm(clause.Tokens[1]),
-                matrix => matrix == null ? "Could not resolve this to a matrix." : null),
+            "material" => GetMaterialInstruction<TObject>(clause),
+            "transform" => GetTransformInstruction<TObject>(clause),
             "no.shadow" => new SetObjectPropertyInstruction<TObject, bool>(
                 target => target.NoShadow, true),
             _ => throw new Exception($"Internal error: unknown {noun} property found: {field}.")
@@ -668,11 +45,13 @@ public partial class LanguageParser
     /// This is a helper method for creating the right instruction for setting the material
     /// property of a surface.
     /// </summary>
-    /// <param name="token">The token that tells us what type of instruction to create.</param>
+    /// <param name="clause">The clause that tells us what type of instruction to create.</param>
     /// <returns>The proper instruction.</returns>
-    private ObjectInstruction<TObject> GetMaterialInstruction<TObject>(Token token)
+    private ObjectInstruction<TObject> GetMaterialInstruction<TObject>(Clause clause)
         where TObject : Surface, new()
     {
+        Token token = clause.Tokens[1];
+
         if (BounderToken.OpenBrace.Matches(token))
         {
             return new SetChildInstruction<TObject, Material>(
@@ -681,7 +60,8 @@ public partial class LanguageParser
 
         if (token.Text == "inherited")
         {
-            if (_context.ParentSet is not GroupInstructionSet)
+            if (_context.ParentSet is not GroupInstructionSet &&
+                _context.ParentSet is not CsgSurfaceInstructionSet)
             {
                 throw new TokenException("The material cannot be inherited since the surface is not in a group.")
                 {
@@ -692,9 +72,45 @@ public partial class LanguageParser
             return new MarkMaterialForInheritanceInstruction<TObject>();
         }
 
+        if (clause.Tokens.Count > 2)
+        {
+            clause.Tokens.RemoveFirst();
+
+            return new SetChildInstruction<TObject, Material>(
+                DetermineProperInstructionSet<MaterialInstructionSet>(
+                    clause, set => ParseMaterialClause(set), true),
+                target => target.Material);
+        }
+
         return new SetObjectPropertyInstruction<TObject, Material>(
             target => target.Material, new VariableTerm(token),
             material => material == null ? "Could not resolve this to a material." : null);
+    }
+
+    /// <summary>
+    /// This is a helper method for creating the right instruction for setting the transform
+    /// property of a surface.
+    /// </summary>
+    /// <param name="clause">The clause that tells us what type of instruction to create.</param>
+    /// <returns>The proper instruction.</returns>
+    private ObjectInstruction<TObject> GetTransformInstruction<TObject>(Clause clause)
+        where TObject : Surface, new()
+    {
+        Token token = clause.Tokens[1];
+
+        if (clause.Tokens.Count > 2)
+        {
+            clause.Tokens.RemoveFirst();
+
+            return new SetChildInstruction<TObject, Matrix>(
+                DetermineProperInstructionSet<TransformInstructionSet>(
+                    clause, set => ParseTransformClause(set), false),
+                target => target.Transform);
+        }
+
+        return new SetObjectPropertyInstruction<TObject, Matrix>(
+            target => target.Transform, new VariableTerm(token),
+            matrix => matrix == null ? "Could not resolve this to a matrix." : null);
     }
 
     /// <summary>
